+++ src/views/settings/components/MediaProxy.vue
@@ -23,7 +23,7 @@
       <p class="expl">The total time the connection is allowed to read from the remote upstream</p>
     </el-form-item>
     <el-form-item label="Inline content types">
-      <el-select :value="inlineContentTypes" @change="processNestedData($event, 'media_proxy', 'proxy_opts', 'inline_content_types')">
+      <el-select :value="inlineContentTypes" clearable @change="processNestedData($event, 'media_proxy', 'proxy_opts', 'inline_content_types')">
         <el-option :value="true" label="True"/>
         <el-option :value="false" label="False"/>
         <el-option value="whitelistedTypeArray" label="List of whitelisted content types"/>
@@ -37,7 +37,7 @@
       </p>
     </el-form-item>
     <el-form-item v-if="inlineContentTypes === 'whitelistedTypeArray'" label="Whitelisted content types">
-      <el-select :value="whitelistedContentTypes" multiple @change="processNestedData($event, 'media_proxy', 'proxy_opts', 'inline_content_types')">
+      <el-select :value="whitelistedContentTypes" multiple filterable allow-create @change="processNestedData($event, 'media_proxy', 'proxy_opts', 'inline_content_types')">
         <el-option
           v-for="item in whitelistedContentTypesOptions"
           :label="item.label"
@@ -54,7 +54,7 @@
       <el-switch :value="http.follow_redirect" @change="processHttpSettings($event, 'media_proxy', 'proxy_opts', 'http', 'follow_redirect')"/>
     </el-form-item>
     <el-form-item label="Pool">
-      <el-select :value="http.pool" @change="processHttpSettings($event, 'media_proxy', 'proxy_opts', 'http', 'pool')">
+      <el-select :value="http.pool" clearable @change="processHttpSettings($event, 'media_proxy', 'proxy_opts', 'http', 'pool')">
         <el-option
           v-for="item in hackneyPoolsOptions"
           :label="item.label"